DescriptorReferences: Add write_proto_file to all generated message classes
Each proto descriptor has an associated FileDescriptor with a name derived
from the name of the message descriptor. For files that reference other
descriptors, this filename is used with an `import` statement, so it's
important that the saved .proto files actually get saved to a file with the
right name so that the imports work. Since the logic for figuring out what
that file name should be is non-trivial, this function removes the need for
the user to figure that out.

https://github.com/IBM/jtd-to-proto/issues/3

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>