DataclassConverter: Remove support for 3.7 and add 3.11 support
Also remove the PPA builds for tests. Those were added in import-tracker
when the library needed to be aware of which modules were c-compiled and
the answer was different on Ubuntu PPA python installs. That's not an issue
for this project, though, so this will simplify the build matrix going
forward.

Reference: https://github.com/IBM/import-tracker/issues/40

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>